requirements.txt
discord
phasedb
pillow